Database_API/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
dnspython
PyJWT
numpy
opencv-python-headless
scikit-learn==1.3.0
vaderSentiment==3.3.2
